# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AncesTree",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/imports/template": {
      "get": {
        "tags": [
          "import"
        ],
        "summary": "Import Template",
        "description": "The columns an import reads, as a CSV that Excel opens. With example=true, filled in\nwith the fictional test family.",
        "operationId": "import_template",
        "parameters": [
          {
            "name": "example",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Example"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/csv; charset=utf-8": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/preview": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Preview Import",
        "description": "What an import would change, for you to tick, its questions, and what it would leave\nout. The import runs inside the database and is rolled back: nothing is written.",
        "operationId": "preview_import",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_import"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports": {
      "get": {
        "tags": [
          "import"
        ],
        "summary": "List Imports",
        "description": "Earlier imports, of spreadsheets and from copies, newest first.",
        "operationId": "list_imports",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImportSummary"
                  },
                  "type": "array",
                  "title": "Response List Imports"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Run Import",
        "description": "Import the file with these answers and the changes ticked: a backup first, then one\nUndo step.",
        "operationId": "run_import",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_run_import"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDone"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/copy/preview": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Preview Copy",
        "description": "What a copy to edit brings back, compared with what the app knows it started from and\nwith the tree now: each change for you to tick. The file is only read, never run.\nRefused (422) when it isn't a copy to edit (\"not_a_copy\", \"not_to_edit\"), is locked without\nits password (\"locked\"), has another (\"wrong_password\") or is damaged (\"damaged\"); (409)\nwhen it wasn't made here (\"unknown_copy\") or is older than one brought in (\"older_copy\").",
        "operationId": "preview_copy",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_copy"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopyPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/copy": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Run Copy",
        "description": "Bring in the changes ticked: a backup first, then the people, details and links as one\nUndo step, then stories and photos. Take back undoes it all later.",
        "operationId": "run_copy",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_run_copy"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDone"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/{import_id}/report": {
      "get": {
        "tags": [
          "import"
        ],
        "summary": "Import Report",
        "description": "What an import left out or found different, as a CSV.",
        "operationId": "import_report",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/{import_id}/take-back": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Take Back Import",
        "description": "Undo an import: everyone it added goes to the Trash, with their links; its other links\ngo; the details it changed go back, and the people it removed come back.",
        "operationId": "take_back_import",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportTakenBack"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_preview_copy": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "A copy to edit, as it came back (.html)"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "The copy's password, when it's locked. Used to open it, then forgotten."
          },
          "answers": {
            "type": "string",
            "title": "Answers",
            "description": "Your answers so far, as JSON: each question's id -> an option's id",
            "default": "{}"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_preview_copy"
      },
      "Body_preview_import": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "A spreadsheet saved as CSV"
          },
          "answers": {
            "type": "string",
            "title": "Answers",
            "description": "Your answers so far, as JSON: each question's id -> an option's id",
            "default": "{}"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_preview_import"
      },
      "Body_run_copy": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "A copy to edit, as it came back (.html)"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "The copy's password, when it's locked. Used to open it, then forgotten."
          },
          "answers": {
            "type": "string",
            "title": "Answers",
            "description": "Your answers so far, as JSON: each question's id -> an option's id",
            "default": "{}"
          },
          "chosen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen",
            "description": "The changes ticked, as a JSON list of their ids. Left out: those ticked to begin with."
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_run_copy"
      },
      "Body_run_import": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "A spreadsheet saved as CSV"
          },
          "answers": {
            "type": "string",
            "title": "Answers",
            "description": "Your answers so far, as JSON: each question's id -> an option's id",
            "default": "{}"
          },
          "chosen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen",
            "description": "The changes ticked, as a JSON list of their ids. Left out: those ticked to begin with."
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_run_import"
      },
...
      "CopyPreview": {
        "properties": {
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "about": {
            "$ref": "#/components/schemas/CopyReturned"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/ImportQuestion"
            },
            "type": "array",
            "title": "Questions"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/ImportChange"
            },
            "type": "array",
            "title": "Changes"
          },
          "left_out": {
            "items": {
              "$ref": "#/components/schemas/ImportLeftOut"
            },
            "type": "array",
            "title": "Left Out"
          },
          "second_look": {
            "items": {
              "$ref": "#/components/schemas/ImportSecondLook"
            },
            "type": "array",
            "title": "Second Look"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "about",
          "questions",
          "changes",
          "left_out",
          "second_look"
        ],
        "title": "CopyPreview",
        "description": "What a copy to edit brings back, for you to tick. Nothing is written."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportDifference": {
        "properties": {
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "person": {
            "type": "string",
            "format": "uuid",
            "title": "Person"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "column": {
            "type": "string",
            "title": "Column"
          },
          "written": {
            "type": "string",
            "title": "Written"
          },
          "in_tree": {
            "type": "string",
            "title": "In Tree"
          }
        },
        "type": "object",
        "required": [
          "row",
          "person",
          "name",
          "column",
          "written",
          "in_tree"
        ],
        "title": "ImportDifference",
        "description": "A row that is someone already in the tree says something the tree doesn't, and it\nisn't offered as a change: the tree has been changed since the file was exported, so the\ntree's is the newer. For you to look at."
      },
      "ImportDone": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "people": {
            "type": "integer",
            "title": "People"
          },
          "links": {
            "type": "integer",
            "title": "Links"
          },
          "changed": {
            "type": "integer",
            "title": "Changed",
            "default": 0
          },
          "removed": {
            "type": "integer",
            "title": "Removed",
            "default": 0
          },
          "stories": {
            "type": "integer",
            "title": "Stories",
            "default": 0
          },
          "photos": {
            "type": "integer",
            "title": "Photos",
            "default": 0
          },
          "left_out": {
            "type": "integer",
            "title": "Left Out"
          },
          "backup": {
            "type": "string",
            "title": "Backup"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "people",
          "links",
          "left_out",
          "backup"
        ],
        "title": "ImportDone"
      },
      "ImportLeftOut": {
        "properties": {
          "row": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row"
          },
          "column": {
            "type": "string",
            "title": "Column"
          },
          "written": {
            "type": "string",
            "title": "Written"
          },
          "why": {
            "type": "string",
            "title": "Why"
          }
        },
        "type": "object",
        "required": [
          "row",
          "column",
          "written",
          "why"
        ],
        "title": "ImportLeftOut",
        "description": "A value or a link that didn't come in, and why; the rest of the row still did. For a\ncopy: whose it was, in `column`, with no row."
      },
      "ImportOption": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "ImportOption",
        "description": "One answer to a question: \"person:<id>\" (someone in the tree), \"row:<n>\" (a row of the\nfile), \"new\" (someone new) or \"out\" (leave the link out)."
      },
      "ImportPerson": {
        "properties": {
          "row": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "born": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Born"
          },
          "birthplace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthplace"
          },
          "named_in": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Named In"
          }
        },
        "type": "object",
        "required": [
          "row",
          "name",
          "born",
          "birthplace",
          "named_in"
        ],
        "title": "ImportPerson",
        "description": "Someone the import adds."
      },
      "ImportPreview": {
        "properties": {
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "rows": {
            "type": "integer",
            "title": "Rows"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/ImportPerson"
            },
            "type": "array",
            "title": "People"
          },
          "matched": {
            "type": "integer",
            "title": "Matched"
          },
          "links": {
            "type": "integer",
            "title": "Links"
          },
          "links_to_tree": {
            "type": "integer",
            "title": "Links To Tree"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/ImportQuestion"
            },
            "type": "array",
            "title": "Questions"
          },
          "left_out": {
            "items": {
              "$ref": "#/components/schemas/ImportLeftOut"
            },
            "type": "array",
            "title": "Left Out"
          },
          "second_look": {
            "items": {
              "$ref": "#/components/schemas/ImportSecondLook"
            },
            "type": "array",
            "title": "Second Look"
          },
          "differences": {
            "items": {
              "$ref": "#/components/schemas/ImportDifference"
            },
            "type": "array",
            "title": "Differences"
          },
          "columns_not_read": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns Not Read"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/ImportChange"
            },
            "type": "array",
            "title": "Changes"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "rows",
          "people",
          "matched",
          "links",
          "links_to_tree",
          "questions",
          "left_out",
          "second_look",
          "differences",
          "columns_not_read"
        ],
        "title": "ImportPreview"
      },
      "ImportQuestion": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "row": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row"
          },
          "column": {
            "type": "string",
            "title": "Column"
          },
          "written": {
            "type": "string",
            "title": "Written"
          },
          "kind": {
            "type": "string",
            "enum": [
              "same_person",
              "which_person"
            ],
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/ImportOption"
            },
            "type": "array",
            "title": "Options"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "id",
          "row",
          "column",
          "written",
          "kind",
          "text",
          "options",
          "answer"
        ],
        "title": "ImportQuestion"
      },
      "ImportSecondLook": {
        "properties": {
          "row": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "row",
          "message"
        ],
        "title": "ImportSecondLook",
        "description": "Something that came in but is worth a look, e.g. an unlikely age gap."
      },
      "ImportSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "spreadsheet",
              "copy"
            ],
            "title": "Kind",
            "default": "spreadsheet"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "for_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "For Name"
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Imported At"
          },
          "people": {
            "type": "integer",
            "title": "People"
          },
          "people_present": {
            "type": "integer",
            "title": "People Present"
          },
          "links": {
            "type": "integer",
            "title": "Links"
          },
          "changed": {
            "type": "integer",
            "title": "Changed",
            "default": 0
          },
          "removed": {
            "type": "integer",
            "title": "Removed",
            "default": 0
          },
          "stories": {
            "type": "integer",
            "title": "Stories",
            "default": 0
          },
          "photos": {
            "type": "integer",
            "title": "Photos",
            "default": 0
          },
          "left_out": {
            "items": {
              "$ref": "#/components/schemas/ImportLeftOut"
            },
            "type": "array",
            "title": "Left Out"
          },
          "differences": {
            "type": "integer",
            "title": "Differences"
          },
          "second_look": {
            "type": "integer",
            "title": "Second Look"
          },
          "taken_back_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taken Back At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "file_name",
          "imported_at",
          "people",
          "people_present",
          "links",
          "left_out",
          "differences",
          "second_look",
          "taken_back_at"
        ],
        "title": "ImportSummary",
        "description": "An earlier import, for Settings \u2192 Import and What's missing."
      },
      "ImportTakenBack": {
        "properties": {
          "moved": {
            "type": "integer",
            "title": "Moved"
          },
          "gone": {
            "type": "integer",
            "title": "Gone"
          },
          "restored": {
            "type": "integer",
            "title": "Restored",
            "default": 0
          },
          "reverted": {
            "type": "integer",
            "title": "Reverted",
            "default": 0
          },
          "kept": {
            "type": "integer",
            "title": "Kept",
            "default": 0
          },
          "links": {
            "type": "integer",
            "title": "Links",
            "default": 0
          },
          "stories": {
            "type": "integer",
            "title": "Stories",
            "default": 0
          },
          "photos": {
            "type": "integer",
            "title": "Photos",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "moved",
          "gone"
        ],
        "title": "ImportTakenBack"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}